refactor: move authority validations to the publish validations
